A super simple proxy that decorates requests before routing them
This is the start of an effort to refactor the hunger app microservice
to support multiple concurrent models and easier rollouts of model
updates. Ideally we'd like ongoing sessions to always get serviced by
the same runtime recomendation service. At the same time, we'd like a
proxy that can easily support A/B (multivariate) testing with an easy
way of configuring which recommendation models are assigned what
percentage of the sessions.

This commit just starts moving us in that direction by implementing a
minimalistic proxy that decorates requests with a searchSession query
string parameter (if one isn't already present) and then routes them to
the actual hunger recommendation service.

The next step will be building it out to support a configuration that
allows support for consistently sending searchSession ids to the same
weighted set of external services.